cleanup

+++ fesvr/htif_zedboard.h
@@ -9,7 +9,7 @@ class htif_zedboard_t : public htif_t
  public:
   htif_zedboard_t(const std::vector<std::string>& args);
   ~htif_zedboard_t();
-  uintptr_t get_host_clk_freq();
+  float get_host_clk_freq();
 
  protected:
   ssize_t read(void* buf, size_t max_size);
